restructure code to make it easier to read
