+++ pkg/util/testappender/internal/dtobuilder/sort.go
@@ -1,4 +1,4 @@
-package testappender
+package dtobuilder
 
 import (
 	"sort"
